// src/types/types.ts
enum Iso639LanguageCode {
    Abkhazian = 'ab',
    Afar = 'aa',
    Afrikaans = 'af',
    Akan = 'ak',
    Albanian = 'sq',
    Amharic = 'am',
    Arabic = 'ar',
    Aragonese = 'an',
    Armenian = 'hy',
    Assamese = 'as',
    Avaric = 'av',
    Avestan = 'ae',
    Aymara = 'ay',
    Azerbaijani = 'az',
    Bambara = 'bm',
    Bashkir = 'ba',
    Basque = 'eu',
    Belarusian = 'be',
    Bengali = 'bn',
    Bihari = 'bh',
    Bislama = 'bi',
    Bosnian = 'bs',
    Breton = 'br',
    Bulgarian = 'bg',
    Burmese = 'my',
    Catalan = 'ca',
    Chamorro = 'ch',
    Chechen = 'ce',
    Chichewa = 'ny',
    Chinese = 'zh',
    Chuvash = 'cv',
    Cornish = 'kw',
    Corsican = 'co',
    Cree = 'cr',
    Croatian = 'hr',
    Czech = 'cs',
    Danish = 'da',
    Divehi = 'dv',
    Dutch = 'nl',
    Dzongkha = 'dz',
    English = 'en',
    Esperanto = 'eo',
    Estonian = 'et',
    Ewe = 'ee',
    Faroese = 'fo',
    Fijian = 'fj',
    Finnish = 'fi',
    French = 'fr',
    Fula = 'ff',
    Galician = 'gl',
    Georgian = 'ka',
    German = 'de',
    Greek = 'el',
    Guarani = 'gn',
    Gujarati = 'gu',
    Haitian = 'ht',
    Hausa = 'ha',
    Hebrew = 'he',
    Herero = 'hz',
    Hindi = 'hi',
    HiriMotu = 'ho',
    Hungarian = 'hu',
    Icelandic = 'is',
    Ido = 'io',
    Igbo = 'ig',
    Indonesian = 'id',
    Interlingua = 'ia',
    Interlingue = 'ie',
    Inuktitut = 'iu',
    Inupiaq = 'ik',
    Irish = 'ga',
    Italian = 'it',
    Japanese = 'ja',
    Javanese = 'jv',
    Kalaallisut = 'kl',
    Kannada = 'kn',
    Kanuri = 'kr',
    Kashmiri = 'ks',
    Kazakh = 'kk',
    Khmer = 'km',
    Kikuyu = 'ki',
    Kinyarwanda = 'rw',
    Kirundi = 'rn',
    Komi = 'kv',
    Kongo = 'kg',
    Korean = 'ko',
    Kurdish = 'ku',
    Kwanyama = 'kj',
    Kyrgyz = 'ky',
    Lao = 'lo',
    Latin = 'la',
    Latvian = 'lv',
    Limburgish = 'li',
    Lingala = 'ln',
    Lithuanian = 'lt',
    LubaKatanga = 'lu',
    Luxembourgish = 'lb',
    Macedonian = 'mk',
    Malagasy = 'mg',
    Malay = 'ms',
    Malayalam = 'ml',
    Maltese = 'mt',
    Manx = 'gv',
    Maori = 'mi',
    Marathi = 'mr',
    Marshallese = 'mh',
    Mongolian = 'mn',
    Nauru = 'na',
    Navajo = 'nv',
    Ndonga = 'ng',
    Nepali = 'ne',
    NorthNdebele = 'nd',
    NorthernSami = 'se',
    Norwegian = 'no',
    NorwegianBokmal = 'nb',
    NorwegianNynorsk = 'nn',
    Occitan = 'oc',
    Ojibwe = 'oj',
    Oriya = 'or',
    Oromo = 'om',
    Ossetian = 'os',
    Pali = 'pi',
    Pashto = 'ps',
    Persian = 'fa',
    Polish = 'pl',
    Portuguese = 'pt',
    Punjabi = 'pa',
    Quechua = 'qu',
    Romanian = 'ro',
    Romansh = 'rm',
    Russian = 'ru',
    Samoan = 'sm',
    Sango = 'sg',
    Sanskrit = 'sa',
    Sardinian = 'sc',
    Serbian = 'sr',
    Shona = 'sn',
    Sindhi = 'sd',
    Sinhala = 'si',
    Slovak = 'sk',
    Slovenian = 'sl',
    Somali = 'so',
    SouthNdebele = 'nr',
    SouthernSotho = 'st',
    Spanish = 'es',
    Sundanese = 'su',
    Swahili = 'sw',
    Swati = 'ss',
    Swedish = 'sv',
    Tagalog = 'tl',
    Tahitian = 'ty',
    Tajik = 'tg',
    Tamil = 'ta',
    Tatar = 'tt',
    Telugu = 'te',
    Thai = 'th',
    Tibetan = 'bo',
    Tigrinya = 'ti',
    Tonga = 'to',
    Tsonga = 'ts',
    Tswana = 'tn',
    Turkish = 'tr',
    Turkmen = 'tk',
    Twi = 'tw',
    Ukrainian = 'uk',
    Urdu = 'ur',
    Uyghur = 'ug',
    Uzbek = 'uz',
    Venda = 've',
    Vietnamese = 'vi',
    Volapuk = 'vo',
    Walloon = 'wa',
    Welsh = 'cy',
    WesternFrisian = 'fy',
    Wolof = 'wo',
    Xhosa = 'xh',
    Yiddish = 'yi',
    Yoruba = 'yo',
    Zhuang = 'za',
    Zulu = 'zu',
}
// ...
enum SourceType {
    EDUCATION = 'education',
    HEALTHCARE = 'healthcare',
    COMPANY = 'company',
    ARCHIEVE = 'archive',
    NONPROFIT = 'nonprofit',
    GOVERNMENT = 'government',
    FACILITY = 'facility',
    OTHER = 'other',
}

enum AuthorPosition {
    FIRST = 'first',
    MIDDLE = 'middle',
    LAST = 'last',
}

type DehydratedInstitution = {
    country_code: Iso3166Alpha2CountryCode;
    display_name: string;
    id: string;
    lineage: string[];
    ror: string;
    type: SourceType;
}

type Affiliation = {
    institution: DehydratedInstitution;
    years: number[];
}

type AuthorCountByYear = {
    year: number;
    works_count?: number;
    cited_by_count: number;
}

type ExternalIds = {
    openalex?: string;
    orcid?: string;
    scopus?: string;
    twitter?: string;
    wikidata?: string;
    issn_l?: string;
    issn?: string[];
    mag?: number;
    fatcat?: string;
}

type SummaryStats = {
    "2yr_mean_citedness": number;
    "h_index": number;
    "i10_index": number;
}

type DehydratedAuthor = {
    id: string;
    display_name: string;
    orcid: string;
}

type DehydratedSource = {
    id: string;
    display_name: string;
    issn_l: string;
    issn: string[];
    host_organization_lineage: string[];
    host_organization_name: string;
    is_core: boolean;
    is_oa: boolean;
    is_in_doaj: boolean;
    type: SourceType;
}
// ...
enum Relationship {
    PARENT = 'parent',
    CHILD = 'child',
    RELATED = 'related',
}

type AssociatedInstitution = {
    id: string;
    ror: string;
    display_name: string;
    country_code: Iso3166Alpha2CountryCode;
    type: SourceType;
    relationship: Relationship;
}

type Geo = {
    city: string;
    geonames_city_id: string;
    region: string;
    country_code: Iso3166Alpha2CountryCode;
    country: string;
    latitude: number;
    longitude: number;
}

type International = {
    display_name: Record<Iso639LanguageCode, string>
}

enum PossibleInstitutionRole {
    FUNDER = 'funder',
    INSTITUTION = 'institution',
    PUBLISHER = 'publisher',
}

type InstitutionRole = {
    role: PossibleInstitutionRole;
    id: string;
    work_count: number;
}

type InstitutionResponse = DehydratedInstitution & {
    associated_institutions: AssociatedInstitution[];
    cited_by_count: number;
    counts_by_year: AuthorCountByYear[];
    created_date: string;
    display_name_acronyms: string[];
    display_name_alternatives: string[];
    geo: Geo;
    homepage_url: string;
    ids: ExternalIds;
    image_thumbnail_url: string;
    is_super_system: boolean;
    image_url: string;
    international: International;
    repositories: SourceResponse[];
    roles: InstitutionRole[];
    summary_stats: SummaryStats;
    updated_date: string;
    works_api_url: string;
    works_count: number;
}

type APCPrice = {
    price: number;
    currency: string;
}

type Society = {
    url: string;
    organization: string;
}

type SourceResponse = DehydratedSource & {
    abbreviated_title: string;
    alternate_titles: string[];
    apc_prices: APCPrice[];
    apc_usd?: number;
    cited_by_count: number;
    country_code: Iso3166Alpha2CountryCode;
    counts_by_year: AuthorCountByYear[];
    created_date: string;
    homepage_url: string;
    host_organization: {
        id: string;
    }
    ids?: ExternalIds;
    societies: Society[];
    summary_stats: SummaryStats;
    updated_date: string;
    works_api_url: string;
    works_count: number;
}